requirements.txt
llama-index-vector-stores-pinecone
google-api-python-client
langchain-community
langchain-pinecone
python-multipart
langchain-openai
pydantic[email]
python-dotenv
llama-index
llama-parse
pdfplumber
langgraph
pinecone
uvicorn
fastapi
twilio
mangum
pytz